Store other links in DB

+++ src/db.rs
@@ -8,19 +8,21 @@ use self::postgres::types::ToSql;
 use self::postgres::rows::Rows;
 use self::postgres::error;
 
+use domain::domain_error::DomainError;
+
 pub struct DB {
-    conn: Option<Connection>,
+    pub conn: Option<Connection>,
 }
 
 impl DB {
-    pub fn new(conn_string: &str) -> DB {
+    pub fn new(conn_string: &str) -> Result<DB, error::ConnectError> {
         match Connection::connect(conn_string, SslMode::None) {
             Ok(conn) => {
-                DB { conn: Some(conn), }
+                Ok(DB { conn: Some(conn), })
             },
             Err(err) => {
                 format!("DB connection error: \n{}", err.description());
-                DB { conn: None, }
+                Err(err)
             }
         }
     }
@@ -30,7 +32,7 @@ impl DB {
             Some(ref c) => c.query(query, params),
             None => {
                 Err(error::Error::Io(io::Error::new(io::ErrorKind::NotConnected,
-                                                    "We are not connected to DB")))
+                                                    "DB: We are not connected to DB")))
             }
         }
     }
@@ -40,7 +42,30 @@ impl DB {
             Some(ref c) => c.execute(query, params),
             None => {
                 Err(error::Error::Io(io::Error::new(io::ErrorKind::NotConnected,
-                                                    "We are not connected to DB")))
+                                                    "DB: We are not connected to DB")))
+            }
+        }
+    }
+
+    pub fn store_domains(&self, list: &Vec<String>) -> Result<u64, error::Error> {
+        let mut modified = 0;
+        match self.conn {
+            Some(ref c) => {
+                match c.prepare("INSERT INTO domain_list (domain_url) VALUES ($1)") {
+                    Ok(stmt) => {
+                        list.iter().map(|url| {
+                            stmt.execute(&[url])
+                                .map_err(|err| err)
+                                .and_then(|res| Ok(modified += res))
+                        }).collect::<Vec<_>>();
+                        Ok(modified)
+                    },
+                    Err(err) => Err(err)
+                }
+            },
+            None => {
+                Err(error::Error::Io(io::Error::new(io::ErrorKind::NotConnected,
+                                                    "DB: We are not connected to DB")))
             }
         }
     }
@@ -49,41 +74,64 @@ impl DB {
 #[cfg(test)]
 #[test]
 fn test_db_connection() {
+    
     let db = DB::new("postgresql://mokosza:mokoszamokosza@\
                       catdamnit.chs4hglw5opg.eu-west-1.rds.amazonaws.com:5432/mokosza");
-
-    assert!(db.conn.is_some());
+    if let Ok(c) = db {
+        assert!(c.conn.is_some());
+    }
 }
 
 #[test]
 fn test_db_query() {
-    let db = DB::new("postgresql://mokosza:mokoszamokosza@\
+    let db_new = DB::new("postgresql://mokosza:mokoszamokosza@\
                       catdamnit.chs4hglw5opg.eu-west-1.rds.amazonaws.com:5432/mokosza");
 
-    assert!(db.conn.is_some());
+    if let Ok(db) = db_new {
+        assert!(db.conn.is_some());
 
-    let _ = db.execute("DROP TABLE test;", &[]);
-    let res = db.execute("CREATE TABLE test (\
-                        id              SERIAL PRIMARY KEY,\
-                        col1            VARCHAR NOT NULL,\
-                        col2            BIGINT\
-                        )", &[]);
-    assert_eq!(res.unwrap(), 0u64);
-
-    let res = db.execute("INSERT INTO test(col1, col2) VALUES($1, $2)", &[&"test", &123456i64]);
-    assert_eq!(res.unwrap(), 1u64);
-
-    let rows = db.query("SELECT id, col1, col2 FROM test WHERE col1 = $1", &[&"test"]);
-    let rows = rows.unwrap();
-    for row in rows.iter() {
-        let id: i32 = row.get("id");
-        let col1: String = row.get(1);
-        let col2: i64 = row.get("col2");
-        assert_eq!(id, 1);
-        assert_eq!(col1, "test");
-        assert_eq!(col2, 123456);
+        let _ = db.execute("DROP TABLE test;", &[]);
+        let res = db.execute("CREATE TABLE test (\
+                              id              SERIAL PRIMARY KEY,\
+                              col1            VARCHAR NOT NULL,\
+                              col2            BIGINT\
+                              )", &[]);
+        assert_eq!(res.unwrap(), 0u64);
+        
+        let res = db.execute("INSERT INTO test(col1, col2) VALUES($1, $2)",
+                             &[&"test", &123456i64]);
+        assert_eq!(res.unwrap(), 1u64);
+        
+        let rows = db.query("SELECT id, col1, col2 FROM test WHERE col1 = $1",
+                            &[&"test"]);
+        let rows = rows.unwrap();
+        for row in rows.iter() {
+            let id: i32 = row.get("id");
+            let col1: String = row.get(1);
+            let col2: i64 = row.get("col2");
+            assert_eq!(id, 1);
+            assert_eq!(col1, "test");
+            assert_eq!(col2, 123456);
+        }
+        
+        let res = db.execute("DROP TABLE test;", &[]);
+        assert_eq!(res.unwrap(), 0u64);
+        
+        let mut list: Vec<String> = Vec::new();
+        list.push("www.example1.com".to_string());
+        list.push("www.example2.com".to_string());
+        list.push("www.example3.com".to_string());
+        let res = db.store_domains(&list);
+        assert_eq!(res.unwrap(), 3u64);
+        
+        let res = db.execute("DELETE FROM domain_list WHERE domain_url='www.example1.com'",
+                             &[]);
+        assert_eq!(res.unwrap(), 1u64);
+        let res = db.execute("DELETE FROM domain_list WHERE domain_url='www.example2.com'",
+                             &[]);
+        assert_eq!(res.unwrap(), 1u64);
+        let res = db.execute("DELETE FROM domain_list WHERE domain_url='www.example3.com'",
+                             &[]);
+        assert_eq!(res.unwrap(), 1u64);
     }
-    
-    let res = db.execute("DROP TABLE test;", &[]);
-    assert_eq!(res.unwrap(), 0u64);
 }
